backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-pptx==0.6.23
moviepy==1.0.3
openai==1.52.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0